Moved isLastLineWhitespaceOnly to general.js and added tests
